requirements.txt
streamlit
pymupdf4llm
langchain
langchain-community
langchain-classic
langchain-text-splitters
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai